refactor: 도메인 개념 단위로 패키지 재구성
- domain.board: Board, Pieces, Position, PositionLayout, Col, Row 통합
- domain.piece: Piece를 game에서 이동, PieceType/Team을 vo에서 이동
- domain.setup: Command를 command에서 이동, Coordinate를 vo에서 이동
- domain.movement: Delta/ColDelta/RowDelta를 vo에서 이동
- domain.game: JanggiGame, Turn만 유지
- domain.vo, domain.command 패키지 제거
