+++ src/main/java/domain/board/Row.java
@@ -1,4 +1,4 @@
-package domain.vo;
+package domain.board;
 
 import java.util.Arrays;
