+++ src/main/java/domain/movement/ColDelta.java
@@ -1,4 +1,4 @@
-package domain.vo;
+package domain.movement;
 
 public record ColDelta(int value) {
 }
