+++ src/main/java/domain/movement/Delta.java
@@ -1,4 +1,4 @@
-package domain.vo;
+package domain.movement;
 
 public record Delta(ColDelta columnDelta, RowDelta rowDelta) {
 }
