+++ src/main/java/domain/movement/RowDelta.java
@@ -1,4 +1,4 @@
-package domain.vo;
+package domain.movement;
 
 public record RowDelta(int value) {
 }
